inventorymanager/requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.8.0
Django==5.2.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.2.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
